requirements.txt
python-dateutil
pytz
django
gunicorn
django-heroku
psycopg2-binary
dj-database-url
django-salesforce
ortools
googlemaps
humanize